Fix: TypeError null is not an object
Sentry captured an issue in our Laravel app

```
TypeError

null is not an object (evaluating 'e.parentNode.querySelectorAll')
```

This probably has to do with the way Inertia.js handles page navigation. My understanding is it replaces the contentes of the <body> entirely. This would likely destroy the `cookies.parentNode` element before this setTimeout is completed.

Perhaps an alternative would be to just call document.querySelectorAll(['data-cookie-consent']) but that fells like a bigger change then I'm comfortable suggestion for this project.